made the file responsive
The media query checks if the screen width is less than or equal to 450px, and if it is, it changes the layout of the header container to have its contents displayed vertically, aligned to the center. The title is also centered and moved to the bottom of the container, and the score container is moved to the top of the container. These styles will be applied when the screen size is smaller than 450px, and the default styles will be applied when the screen size is greater than 450px.